Added overloads for Server/ClientPacketID, support for the MessageProcessor to read a defined number of bits for the ID, and the number of bits required discovered at runtime via reflection.

+++ netgore/trunk/DemoGame.Client/Networking/ClientPacket.cs
@@ -44,7 +44,7 @@ namespace DemoGame.Client
         static PacketWriter GetWriter(ClientPacketID id)
         {
             PacketWriter pw = _writerPool.Create();
-            pw.Write((byte)id);
+            pw.Write(id);
             return pw;
         }
